# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ManualPunch:
    properties:
      approved_at:
        type: string
      approver_id:
        description: References employees(id) - who approved/rejected
        type: integer
      approver_name:
        type: string
      created_at:
        description: Timestamps
        type: string
      employee_id:
        description: References employees(id)
        type: integer
      employee_name:
        description: Virtual fields (not stored in database)
        type: string
      id:
        type: integer
      punch_date:
        description: Date of the punch
        type: string
      punch_time:
        description: Actual time of punch
        type: string
      punch_type:
        allOf:
        - $ref: '#/definitions/models.ManualPunchType'
        description: in or out
      reason:
        description: Reason for manual punch
        type: string
      rejection_reason:
        description: Reason for rejection if applicable
        type: string
      status:
        $ref: '#/definitions/models.ManualPunchStatus'
      supporting_document:
        description: Optional document URL
        type: string
      updated_at:
        type: string
    type: object
  models.ManualPunchApprovalRequest:
    properties:
      rejection_reason:
        description: Required if status is rejected
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ManualPunchStatus'
        enum:
        - approved
        - rejected
    required:
    - status
    type: object
  models.ManualPunchRequest:
    properties:
      punch_type:
        allOf:
        - $ref: '#/definitions/models.ManualPunchType'
        enum:
        - in
        - out
      reason:
        maxLength: 500
        minLength: 10
        type: string
      supporting_document:
        type: string
    required:
    - punch_type
    - reason
    type: object
  models.ManualPunchStatus:
    enum:
    - pending
    - approved
    - rejected
    - cancelled
    type: string
    x-enum-varnames:
    - ManualPunchStatusPending
    - ManualPunchStatusApproved
    - ManualPunchStatusRejected
    - ManualPunchStatusCancelled
  models.ManualPunchType:
    enum:
    - in
    - out
    type: string
    x-enum-varnames:
    - ManualPunchTypeIn
    - ManualPunchTypeOut
...
  response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hrms.com
    name: API Support
    url: https://github.com/Josiahmpokera-dev/hrms-backend
  description: Human Resource Management System - REST API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: HRMS Backend API
  version: "1.0"
paths:
...
  /attendance/manual-punches:
    get:
      consumes:
      - application/json
      description: Get all manual punch requests with optional filtering and pagination
      parameters:
      - description: Filter by employee ID
        in: query
        name: employee_id
        type: integer
      - description: Filter by status (pending, approved, rejected, cancelled)
        in: query
        name: status
        type: string
      - description: Filter by punch type (in, out)
        in: query
        name: punch_type
        type: string
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Manual punches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ManualPunch'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all manual punch requests (Admin/HR)
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      description: Create a manual attendance punch request for when biometric/fingerprint
        fails
      parameters:
      - description: Manual punch request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ManualPunchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Manual punch created successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a manual punch request
      tags:
      - Attendance
  /attendance/manual-punches/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a manual punch request (soft delete)
      parameters:
      - description: Manual punch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Manual punch deleted successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Manual punch not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete manual punch request (Admin/HR)
      tags:
      - Attendance
    get:
      consumes:
      - application/json
      description: Get a specific manual punch request by ID
      parameters:
      - description: Manual punch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Manual punch retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ManualPunch'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Manual punch not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get manual punch by ID
      tags:
      - Attendance
  /attendance/manual-punches/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a pending manual punch request
      parameters:
      - description: Manual punch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Manual punch cancelled successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Manual punch not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Cancel manual punch request
      tags:
      - Attendance
  /attendance/manual-punches/{id}/status:
    patch:
      consumes:
      - application/json
      description: Approve or reject a manual punch request
      parameters:
      - description: Manual punch ID
        in: path
        name: id
        required: true
        type: integer
      - description: Approval decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ManualPunchApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Manual punch status updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ManualPunch'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Manual punch not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve/Reject manual punch request (Admin/HR)
      tags:
      - Attendance
  /attendance/manual-punches/my:
    get:
      consumes:
      - application/json
      description: Get all manual punch requests for the authenticated employee
      parameters:
      - description: Filter by status (pending, approved, rejected, cancelled)
        in: query
        name: status
        type: string
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Manual punches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ManualPunch'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get my manual punch requests
      tags:
      - Attendance
  /attendance/manual-punches/pending:
    get:
      consumes:
      - application/json
      description: Get all pending manual punch requests for approval
      produces:
      - application/json
      responses:
        "200":
          description: Pending manual punches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ManualPunch'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get pending manual punch requests (Admin/HR)
      tags:
      - Attendance
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"